Changing a default heading from h3 to h2
The web typography update modified the standardized practice of which
headings (H1, H2, etc) are used on templated pages.

However, the default "lorem ipsum" text on page templates uses H3 as the
default instead of the new H2. For users who don't modify it themselves,
this means that the headings become inconsistent on published pages,
inhibiting accessibility and best practices.

The lorem ipsum headings need to be modified to H2 in the TMPL files.